Fix bug with declaredMethods in gumtree building Rename package `preprocessing` to `gumtree`
