+++ src/main/kotlin/org/jetbrains/research/gumtree/PyPSIGumTree.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.research.preprocessing
+package org.jetbrains.research.gumtree
 
 import com.github.gumtreediff.tree.Tree
 import com.jetbrains.python.psi.PyElement
@@ -6,7 +6,10 @@ import org.jetbrains.research.plugin.pyflowgraph.getFullName
 import org.jetbrains.research.plugin.pyflowgraph.getType
 
 
-class PyPSIGumTree(private val rootElement: PyElement) : Tree(rootElement.getType(), rootElement.getFullName()) {
+class PyPSIGumTree(val rootElement: PyElement) : Tree(
+    rootElement.getType(),
+    rootElement.getFullName()
+) {
 
     private val fieldNames: MutableList<String> = arrayListOf()
